codec and validation are now in their own separate packages
