+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [workspace]
-members = ["wf_common"]
+members = ["wf_common", "wf_validation", "wf_codec"]
 
 
 [dependencies]
@@ -20,3 +20,5 @@ hkdf = "0.12.3"
 sha2 = "0.10.2"
 lazy_static = "1.4.0"
 wf_common = { path = "wf_common" }
+wf_validation = { path = "wf_validation" }
+wf_codec = { path = "wf_codec" }
