+++ src/lib.rs
@@ -7,14 +7,12 @@ mod error;
 mod wf_account;
 mod wf_auth;
 mod wf_buffer;
-mod wf_codec;
 #[allow(dead_code)]
 mod wf_core;
 mod wf_crypto;
 mod wf_field;
 mod wf_json;
 mod wf_parser;
-mod wf_validation;
 
 pub fn encode_from_json<T: AsRef<str>>(json: T) -> Result<String, WhiteflagError> {
     let message: wf_json::WhiteflagFieldValues =
