--- src/wf_codec/mod.rs
@@ -1,4 +1,0 @@
-pub mod binary;
-pub mod encoding;
-pub mod hexadecimal;
-pub mod latlong;
