+++ wf_codec/src/binary.rs
