+++ wf_codec/src/hexadecimal.rs
