+++ wf_codec/src/lib.rs
@@ -1,0 +1,10 @@
+#[macro_use]
+extern crate lazy_static;
+
+pub mod binary;
+pub mod encoding;
+pub mod hexadecimal;
+pub mod latlong;
+
+#[cfg(test)]
+mod validation_tests;
