+++ wf_validation/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "wf_validation"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0.31"
