+++ wf_validation/src/lib.rs
@@ -1,9 +1,3 @@
-use crate::wf_codec::encoding::Encoding;
-use crate::wf_field::FieldDefinition;
-
-#[cfg(test)]
-mod test;
-
 /// there are four ways a value can be resolved as invalid
 /// - [Encoding] invalid length according to encoding
 /// - [Encoding] invalid character set according to encoding
@@ -25,20 +19,31 @@ pub trait Validation {
     fn validate(&self, value: &str) -> Result<(), ValidationError>;
 }
 
-const NULL_FIELD_NAME: &'static str = "NULL FIELD NAME";
-
-impl Validation for FieldDefinition {
-    fn validate(&self, value: &str) -> Result<(), ValidationError> {
-        match self.expected_byte_length() {
-            Some(len) if len != value.len() => Err(ValidationError::InvalidLength {
-                data: value.to_string(),
-                expected_length: len,
-                specification_level: format!(
-                    "== Field Definition Error for {} ==",
-                    self.get_name().unwrap_or(NULL_FIELD_NAME)
-                ),
-            }),
-            _ => self.encoding.validate(value),
-        }
+pub fn test<T: Validation>(validator: T, data: &str) -> Result<(), ValidationError> {
+    let result = validator.validate(data);
+    assert!(result.is_err(), "the test did not throw an error");
+    if result.is_err() {
+        println!("{}", result.as_ref().unwrap_err());
     }
+    result
+}
+
+pub fn invalid_length_test<T: Validation>(validator: T, data: &str, expected_length: usize) {
+    // ensure we have an incorrect length
+    assert!(
+        data.len() != expected_length,
+        "the input data byte_length should not be correct"
+    );
+
+    assert!(
+        matches!(
+            test(validator, data),
+            Err(ValidationError::InvalidLength {
+                data: _,
+                expected_length: _,
+                specification_level: _
+            })
+        ),
+        "InvalidLength was not thrown"
+    );
 }
